Sketch Redis offest backend
